pyproject.toml
[tool.poetry]
name = "lms-api-mvp"
version = "0.1.0"
description = ""
authors = ["aahnik <daw@aahnik.dev>"]
readme = "README.md"
packages = [{include = "api"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.0"
beanie = "^1.17.0"
uvicorn = "^0.21.1"
python-decouple = "^3.8"
bcrypt = "^4.0.1"
pydantic = {extras = ["email"], version = "^1.10.6"}
passlib = "^1.7.4"
python-jose = "^3.3.0"
python-multipart = "^0.0.6"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"
ruff = "^0.0.257"
jupyter = "^1.0.0"
notebook = "^6.5.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''
